Move backtesting api to it's own file
